Update test-gen-support from master

+++ src/ansi-c/ansi_c_scope.cpp
@@ -29,7 +29,7 @@ void ansi_c_scopet::print(std::ostream &out) const
   for(const auto &name : name_map)
   {
     out << "  ID: " << name.first
-        << " CLASS: " << name.second.id_class
+        << " CLASS: " << static_cast<int>(name.second.id_class)
         << "\n";
   }
 }
